refactor(GroupedList): rename component to VirtualGroupedList
